move StatusCodeError out of botFrameworkAdapter.ts
