+++ libraries/botbuilder/src/index.ts
@@ -11,7 +11,6 @@ export {
     BotFrameworkAdapterSettings,
     INVOKE_RESPONSE_KEY,
     StatusCodes,
-    StatusCodeError,
 } from './botFrameworkAdapter';
 export { BotFrameworkHttpClient } from './botFrameworkHttpClient';
 export { ChannelServiceHandler } from './channelServiceHandler';
@@ -23,6 +22,7 @@ export {
     WebResponse
 } from './interfaces';
 export * from './skills';
+export { StatusCodeError } from './statusCodeError';
 export { StreamingHttpClient, TokenResolver } from './streaming';
 export * from './teamsActivityHandler';
 export * from './teamsActivityHelpers';
