+++ libraries/botbuilder/src/statusCodeError.ts
@@ -1,0 +1,17 @@
+/**
+ * @module botbuilder
+ */
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { StatusCodes } from './botFrameworkAdapter';
+
+export class StatusCodeError extends Error {
+    public readonly statusCode: StatusCodes;
+    public constructor(statusCode: StatusCodes, message?: string) {
+        super(message);
+        this.statusCode = statusCode;
+    }
+}
